requirements.txt
selenium==4.11.0
python-telegram-bot==20.3
python-dotenv==1.0.0
psutil==5.9.5
chromedriver-autoinstaller